+++ plone/app/vocabularies/tests/test_imagesvocabularies.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from plone.app.vocabularies.testing import PAVocabularies_INTEGRATION_TESTING
-from plone.registry.interfaces import IRegistry
 from zope.component import getUtility
 from zope.schema.interfaces import IVocabularyFactory
 import unittest