+++ plone/app/vocabularies/tests/test_imagesvocabularies.py
@@ -2,6 +2,7 @@
 from plone.app.vocabularies.testing import PAVocabularies_INTEGRATION_TESTING
 from zope.component import getUtility
 from zope.schema.interfaces import IVocabularyFactory
+
 import unittest