+++ plone/app/vocabularies/tests/test_subjects_under_context.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
+from plone.app.layout.navigation.interfaces import INavigationRoot
 from plone.app.vocabularies.testing import PAVocabularies_INTEGRATION_TESTING
 from zope.interface import alsoProvides
-from plone.app.layout.navigation.interfaces import INavigationRoot
 
 import mock
 import unittest2 as unittest