+++ plone/app/vocabularies/tests/test_timezonevocabularies.py
@@ -3,6 +3,7 @@ from plone.app.vocabularies.testing import PAVocabularies_INTEGRATION_TESTING
 from plone.registry.interfaces import IRegistry
 from zope.component import getUtility
 from zope.schema.interfaces import IVocabularyFactory
+
 import unittest2 as unittest