+++ CHANGES.rst
@@ -14,8 +14,8 @@ New features:
 
 Bug fixes:
 
-- *add item here*
-
+- Adapt code to follow Plone styleguide.
+  [gforcada]
 
 3.0 (2016-06-12)
 ----------------